Fix compiler warning
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1495 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/lib/abyss/src/socket_unix.c
@@ -304,7 +304,7 @@ channelWait(TChannel * const channelP,
     pollfds[1].events = POLLIN;
     
     rc = poll(pollfds, ARRAY_SIZE(pollfds),
-              timeoutMs == TIME_INFINITE ? -1 : timeoutMs);
+              timeoutMs == TIME_INFINITE ? -1 : (int)timeoutMs);
 
     if (rc < 0) {
         if (errno == EINTR) {
